+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - ActivityDryForm
 - Activities page for guilds
 - creating activities
+- activities page with last guild news
 
 ## [1.1.9] - 2019-12-24
 ### Modified